hw1/requirements.txt
pycryptodome == 3.19.0